+++ src/tile/tile_cache.ts
@@ -210,11 +210,11 @@ export class TileCache {
         }
     }
     abortAllRequests() {
-        for (const key in this.data) {
-            for (const entry of this.data[key]) {
-                if (entry.value && entry.value.abortController) {
-                    entry.value.abortController.abort();
-                }
+        for (const [, entries] of Object.entries(this.data)) {
+            for (const entry of entries) {
+                if (!entry.value) continue;
+                entry.value.aborted = true;
+                entry.value.abortController?.abort();
             }
         }
     }